+++ lib/generators/cms/component/contact_page/contact_page_generator.rb
@@ -1,23 +1,16 @@
-require 'generators/cms/migration'
-
 module Cms
   module Generators
     module Component
       class ContactPageGenerator < ::Rails::Generators::Base
         include Migration
+        include BasePaths
 
         class_option :homepage_path,
           type: :string,
-          default: '/website/de',
+          default: nil,
           desc: 'Path to a CMS homepage, for which to create the contact form.'
 
         source_root File.expand_path('../templates', __FILE__)
-
-        def copy_app_directory
-          directory('app')
-          directory('config')
-          directory('spec')
-        end
 
         def add_email_validation
           gem('valid_email', '0.0.4')
@@ -72,15 +65,47 @@ module Cms
         end
 
         def create_migration
-          validate_obj_class(class_name)
-          validate_attribute(crm_activity_type_attribute_name)
+          begin
+            validate_attribute(crm_activity_type_attribute_name)
+            Rails::Generators.invoke('cms:attribute', [crm_activity_type_attribute_name, '--type=string', '--title=CRM Activity Type'])
+          rescue DuplicateResourceError
+          end
 
-          migration_template('migration.rb', 'cms/migrate/create_contact_page.rb')
+          begin
+            validate_attribute(redirect_after_submit_attribute_name)
+            Rails::Generators.invoke('cms:attribute', [redirect_after_submit_attribute_name, '--type=linklist', '--title=Redirect After Submit Page', '--max-size=1'])
+          rescue DuplicateResourceError
+          end
+
+          begin
+            validate_attribute(contact_page_attribute_name)
+            Rails::Generators.invoke('cms:attribute', [contact_page_attribute_name, '--type=linklist', '--title=Contact Page', '--max-size=1'])
+          rescue DuplicateResourceError
+          end
+
+          begin
+            validate_attribute(show_in_navigation_attribute_name)
+            Rails::Generators.invoke('cms:attribute', [show_in_navigation_attribute_name, '--type=boolean', '--title=Show in Navigation'])
+          rescue DuplicateResourceError
+          end
+
+          begin
+            validate_obj_class(class_name)
+            Rails::Generators.invoke('cms:model', [class_name, '--title=Page: Contact', "--attributes=#{crm_activity_type_attribute_name}", redirect_after_submit_attribute_name, show_in_navigation_attribute_name])
+          rescue DuplicateResourceError
+          end
+
+          migration_template('example_migration.rb', 'cms/migrate/create_contact_page_example.rb')
 
           if behavior == :invoke
             log(:migration, 'Make sure to run "rake cms:migrate" to apply CMS changes and "bundle" to install new gem.')
           end
-        rescue DuplicateResourceError
+        end
+
+        def copy_app_directory
+          directory('app', force: true)
+          directory('config', force: true)
+          directory('spec', force: true)
         end
 
         def remove_custom_type
@@ -101,8 +126,9 @@ module Cms
 
         private
 
+        alias_method :original_homepage_path, :homepage_path
         def homepage_path
-          options[:homepage_path]
+          options[:homepage_path] || original_homepage_path
         end
 
         def class_name
@@ -115,6 +141,14 @@ module Cms
 
         def redirect_after_submit_attribute_name
           'redirect_after_submit_link'
+        end
+
+        def contact_page_attribute_name
+          'contact_page_link'
+        end
+
+        def show_in_navigation_attribute_name
+          'show_in_navigation'
         end
 
         def activity_type