+++ lib/generators/cms/component/google_analytics/templates/migration.rb
@@ -1,94 +1,57 @@
-class CreateGoogleAnalytics < ::RailsConnector::Migrations::Migration
+class IntegrateGoogleAnalytics < ::RailsConnector::Migrations::Migration
   def up
-    create_attributes_and_obj_class
-    create_objs
+    preset_attributes
+    create_configuration_obj
     deactivate_obj_class
     add_hompage_attribute
-    connect_hompages_with_configuration
+    add_homepage_configuration
   end
 
   private
 
-  def class_name
-    '<%= class_name %>'
+  def path
+    '<%= configuration_path %>/google_analytics'
   end
 
-  def obj_path_for_homepage(homepage)
-    "#{homepage.path}/_configuration/google_analytics"
-  end
-
-  def create_attributes_and_obj_class
-    create_attribute(
-      name: '<%= tracking_id_attribute_name %>',
-      title: 'Tracking ID',
-      type: :string
-    )
-
-    create_attribute(
-      name: '<%= anonymize_ip_attribute_name %>',
-      title: 'Anonymize IP?',
-      type: :enum,
-      values: ['Yes', 'No']
-    )
-
-    params = {
-      name: class_name,
-      type: :publication,
-      title: 'Google Analytics',
-      attributes: [
-        '<%= tracking_id_attribute_name %>',
-        '<%= anonymize_ip_attribute_name %>',
-      ],
+  def preset_attributes
+    update_obj_class(
+      '<%= class_name %>',
       preset_attributes: {
         '<%= tracking_id_attribute_name %>' => '<%= tracking_id_default %>',
         '<%= anonymize_ip_attribute_name %>' => '<%= anonymize_ip_default %>',
-      },
-    }
-
-    create_obj_class(params)
+      }
+    )
   end
 
-  def homepages
-    @homepages ||= Obj.find_by_path('/website').toclist
-  end
-
-  def create_objs
-    homepages.each do |homepage|
-      create_obj(
-        _path: obj_path_for_homepage(homepage),
-        _obj_class: class_name,
-        title: 'Google Analytics'
-      )
-    end
+  def create_configuration_obj
+    create_obj(
+      _path: path,
+      _obj_class: '<%= class_name %>',
+      title: 'Google Analytics'
+    )
   end
 
   def deactivate_obj_class
-    update_obj_class(class_name, is_active: false)
+    update_obj_class(
+      '<%= class_name %>',
+      is_active: false
+    )
   end
 
   def add_hompage_attribute
-    create_attribute(
-      name: attribute_name,
-      title: 'Google Analytics',
-      type: :linklist,
-      max_size: 1
+    attributes = get_obj_class('Homepage')['attributes']
+    attributes << '<%= homepage_configuration_attribute_name %>'
+
+    update_obj_class('Homepage', attributes: attributes)
+  end
+
+  def add_homepage_configuration
+    homepage = Obj.find_by_path('<%= homepage_path %>')
+    obj = Obj.find_by_path(path)
+
+    update_obj(
+      homepage.id,
+      '<%= homepage_configuration_attribute_name %>' => [{ obj_id: obj.id }]
     )
-
-    obj_class_attributes = get_obj_class('Homepage')['attributes']
-    obj_class_attributes << attribute_name
-    update_obj_class('Homepage', attributes: obj_class_attributes)
-  end
-
-  def attribute_name
-    'google_analytics'
-  end
-
-  def connect_hompages_with_configuration
-    homepages.each do |homepage|
-      path = obj_path_for_homepage(homepage)
-      obj = Obj.find_by_path(path)
-
-      update_obj(homepage.id, google_analytics: [{ obj_id: obj.id }])
-    end
   end
 end