+++ lib/generators/cms/component/language_switch/language_switch_generator.rb
@@ -1,5 +1,3 @@
-require 'generators/cms/migration'
-
 module Cms
   module Generators
     module Component