+++ lib/ice_kickstarter.rb
@@ -4,6 +4,8 @@ require 'haml'
 require 'ice_kickstarter/engine'
 require 'ice_kickstarter/configuration'
 require 'ice_kickstarter/dashboard'
+require 'generators/cms/migration'
+require 'generators/cms/base_paths'
 
 module IceKickstarter
 end